+++ packages/api/lib/request.js
@@ -1,26 +1,22 @@
-const { RecordDoesNotExist } = require('@cumulus/common/errors');
 const log = require('@cumulus/common/log');
 
-const { User } = require('../models');
 const {
   TokenUnauthorizedUserError
 } = require('../lib/errors');
 const { verifyJwtToken } = require('./token');
+const { isAuthorizedOAuthUser } = require('../app/auth');
 
 /**
  * Verify the validity and access of JWT for request authorization.
  *
  * @param {string} requestJwtToken - The JWT token used for request authorization
- * @param {Object} params - additional parameters
- * @param {string} params.usersTable - The name of the DynamoDB Users table
- *
  * @throws {JsonWebTokenError} - thrown if the JWT is invalid
  * @throws {TokenExpiredError} - thown if the JWT is expired
  * @throws {TokenUnauthorizedUserError} - thrown if the user is not authorized
  *
  * @returns {string} accessToken - The access token from the OAuth provider
  */
-async function verifyJwtAuthorization(requestJwtToken, params = {}) {
+async function verifyJwtAuthorization(requestJwtToken) {
   let accessToken;
   let username;
   try {
@@ -30,13 +26,8 @@ async function verifyJwtAuthorization(requestJwtToken, params = {}) {
     throw err;
   }
 
-  const userModel = new User(params);
-  try {
-    await userModel.get({ userName: username });
-  } catch (err) {
-    if (err instanceof RecordDoesNotExist) {
-      throw new TokenUnauthorizedUserError();
-    }
+  if (!(await isAuthorizedOAuthUser(username))) {
+    throw new TokenUnauthorizedUserError();
   }
 
   return accessToken;