+++ packages/common/aws.js
@@ -383,6 +383,17 @@ exports.getS3Object = improveStackTrace(
     )
 );
 
+exports.getJsonS3Object = (bucket, key) =>
+  exports.getS3Object(bucket, key)
+    .then(({ Body }) => JSON.parse(Body.toString()));
+
+exports.putJsonS3Object = (bucket, key, data) =>
+  exports.s3PutObject({
+    Bucket: bucket,
+    Key: key,
+    Body: JSON.stringify(data)
+  });
+
 exports.getS3ObjectReadStream = (bucket, key) => exports.s3().getObject(
   { Bucket: bucket, Key: key }
 ).createReadStream();